+++ slider/src/app/appBase.h
