+++ slider/src/brain/appConfig.h
