+++ slider/src/brain/settings.cpp
