+++ slider/src/brain/settings.h
